Moving spec_func and obj_index_data out of merc.h
